Minor corrections.

+++ MyTimer/app/src/test/java/fi/fimurito/mytimer/UtilsTest.kt
@@ -11,7 +11,7 @@ class UtilsTest {
     fun testTaskTimer() {
         val n = ZonedDateTime.now()
         val t = Utils.getNextTaskTimer()
-        val m: Int = n.minute - n.minute % AppConstants.CURRENT_MINUTE_DIVISOR
+        val m:  Long = n.minute - n.minute % AppConstants.CURRENT_MINUTE_DIVISOR
 
         // Log.d(AppConstants.LOG_TAG, "time: ${t.hour}:${t.minute}:${t.second}")
